Extra target and files related methods to middleware
